Commit #0003 Add more instructions support to compiler Removed warnings popping up when building Added `rebuild` function to the Makefile

+++ src/lexer.c
@@ -41,7 +41,7 @@ void lexer_tokenize(lexer * lex) {
             token * tok = (token*)malloc(sizeof(token));
             tok->type = NUMBER;
             tok->start = cstart;
-            tok->value = (void*)x;
+            tok->value = (void*)(size_t)x; // (size_t) added here to remove warnings
             tok->line = lindex;
             tokenlist_push(lex->toklist, tok);
         } else if(c == '.') {
